Enum vcf_rider::indel::MutationClass

pub enum MutationClass {
    Manage(usize),
    Ins(Vec<u8>, usize),
    Del(u64, usize),
    Reference,
}

Used to classify indels and snps in groups: SNPs will be tagged as "Manage" while indels with Ins or Del (if this group has their alternative allele). String (Vec) is the sequence that needs to be inserted for ins, usize is the coord inside the window and u64 the length of deletions.

Variants
Manage(usize)Ins(Vec<u8>, usize)Del(u64, usize)Reference