Struct vcf_rider::pwm::Matrix

pub struct Matrix { /* fields omitted */ }

Struct used to represent PWM likelihoods/nucleotide frequencies. ncols is used to be able to represent 4 and 5 nucleotides PWMs (to manage N in sequences). No consistency checks are performed for get/set operations for efficiency.

Methods

impl Matrix

pub fn new(n: usize) -> Matrix

Creates a new matrix with n columns.

Arguments

n - the number of columns

pub fn with_capacity(n: usize, len: usize) -> Matrix

Creates a new matrix preallocated for len elements. Warning: len is not checked to be compatible with n.

Arguments

n - the number of columns

len - the number of elements to be preallocated

pub fn len(&self) -> usize

Returns the number of rows of this Matrix

pub fn push_row_4(&mut self, a: f64, c: f64, g: f64, t: f64)

Adds a 4 element row to this Matrix. No checks of compatibility with self.ncols.

Arguments

a - the likelihood for nucleotide A that we want to add

c - the likelihood for nucleotide C that we want to add

g - the likelihood for nucleotide G that we want to add

t - the likelihood for nucleotide T that we want to add

pub fn push_row_5(&mut self, a: f64, c: f64, g: f64, t: f64, n: f64)

Adds a 5 element row to this Matrix. No checks of compatibility with self.ncols.

Arguments

a - the likelihood for nucleotide A that we want to add

c - the likelihood for nucleotide C that we want to add

g - the likelihood for nucleotide G that we want to add

t - the likelihood for nucleotide T that we want to add

n - the likelihood for nucleotide N that we want to add

pub fn get(&self, row: usize, col: usize) -> f64

Returns the element in row, col of this Matrix. No boundary checks are performed!