Function vcf_rider::rider::encode_genotypes

pub fn encode_genotypes(
    snps_buffer: &VecDeque<Mutation>, 
    overlapping_info: &Vec<(usize, MutationClass)>, 
    group: &Vec<usize>, 
    n_samples: usize, 
    id_samples: &Vec<usize>
) -> Vec<BitVec>

Function that encode the genotypes of our individuals as BitVec given the list of the overlapping mutations, the information on their indel status obtained from IndelRider, the indexes of this group individuals/alleleles, the total number of individuals that we are studying and a vector of all their ids.

Arguments

snps_buffer- a mutable reference to the VecDeque that is used as a buffer for SNPs

overlapping_info - informations about the mutations overlapping with this window/// * reference - the reference genome sequence for the chr where this window falls

group- the ids of the individuals/alleles of this group

n_samples - the total number of individuals in our vcf, needed to distinguish maternal and paternal alleles

id_samples - a vector to map from allele indexes to individual indexes (in genotypes we have the two alleles in tuples .0/.1, while we want to flatten them to our indexes: For individual 1 the two alleles are 0 and n_samples, for individual 2 1 and n_samples+1)