Function vcf_rider::rider::find_overlapping_snps

pub fn find_overlapping_snps<I>(
    window: &Coordinate, 
    reader: &mut I, 
    snps_buffer: &mut VecDeque<Mutation>
) -> usize where
    I: Iterator<Item = Mutation>, 

Function that advances on the VcfReader (Iterator of Mutation) until the first snp that does not overlap with the given window, putting in snps_buffer all the overlapping snps and their number and then the first not overlapping snp. It uses SNPs in snps_buffer to manage overlapping (always sorted on their starting coord!) bed entries. b3_e < b2_e TODO is managed/test?

Arguments

window - the window where we search for overlapping SNPs. This function should be called giving them in order with respect to genomic coords

reader - a mutable reference to an Iterator of Mutation. This will be consumed, it needs to store Mutation objects in order

snps_buffer- a mutable reference to the VecDeque that is used as a buffer for SNPs. SNPs before the given window will be removed, the overlapping ones will be at positions 0..returned value and the first SNPs after the given window will be the last element