Function vcf_rider::rider::get_scores

pub fn get_scores<T: CanScoreSequence>(
    params: RiderParameters<T>, 
    vcf_path: &str, 
    bed_reader: Reader<File>, 
    ref_path: &str, 
    associations: Option<String>, 
    accept_unphased: bool
)

The single entry point of our library, right now for ease of use in bioinformatic pipelines it simply prints the results on standard output. TODO: return a suitable data structure with results.

Arguments

params - the RiderParameter that will be used to score individual sequences

vcf_path - a path to a vcf file representing mutations on a given chr for some individuals

bed_reader - a bed::Reader representing the genomic intervals that we want to consider. They should be on a single chr, sorted and not overlapping. They should not fall outside of the given chromosome sequence

ref_path - the path to a fasta file with the reference sequence of the given chr

associations - an optional string, if given this will be the file name where associations between bed ids and polymorphism will be printed - for bed with at least an overlapping SNPs we will print the ids of the overlapping SNPs (ids are starting coord, indel_length and boolean is_indel - note that insertions are represent with a negative length and deletion with a positive one).

accept_unphased - if unphased vcf files will be accepted. Some scores (i.e. TBA) need phased genotypes cause different phases will end up in different scores, while other (i.e. GC content) are invariant with respect to phasing and for them we will obtain sequences considering even unphased vcf as phased

Panics

If the bed defines some intervals out of the given chromosome fasta. If the given fasta is empty, it's a multifasta, it's not readable or it's not a fasta. It it cannot open the given associations file.