Function vcf_rider::rider::obtain_seq

pub fn obtain_seq(
    window: &Coordinate, 
    snps_buffer: &VecDeque<Mutation>, 
    overlapping_info: &Vec<(usize, MutationClass)>, 
    reference: &Fasta, 
    genotypes: &Vec<BitVec>, 
    seqs: &mut Vec<(BitVec, Vec<u8>)>, 
    bed_end: u64
)

Function that populates a vector of sequences, given a genomic window, a buffer of SNPs overlapping with it alongside their indel status information for the samples groups, the reference sequence, the genotypes encoded for our individuals and the end of the bed that we are considering. The filled vector is a vector of tuples of BitVec, representing the sequences indexes, and Vec, representing the sequences itself. At this stage we will remove duplicated sequences, storing only once the ones that are the same in different individual/alleles.

Arguments

window- the genomic coordinates of the window whose sequences we want to obtain

snps_buffer- a mutable reference to the VecDeque that is used as a buffer for SNPs

overlapping_info - informations about the mutations overlapping with this window

reference - the reference genome sequence for the chr where this window falls

genotypes- a vector of the BitVec indexes of the individuals/alleles that we are studying

seqs - the vector that we will fill with the obtained window sequences

bed_end - the end coordinate of the bed that we are studying, needed to trim the window end if it falls outside of the bed