Struct vcf_rider::rider::RiderParameters

pub struct RiderParameters<'a, T: CanScoreSequence + 'a> {
    pub min_len: usize,
    pub max_len: usize,
    pub parameters: &'a Vec<T>,
}

The parameters used to setup vcf_rider are a vector of objects able to score a sequence and their minimum and maximum lengths. In the future it will become possible to combine scores for all the subsequences not only summing but also for example averaging them, getting the minimum or the maximum, etc.

Fields
min_len: usize

The minimum length of the sequence that these T objects can score. Not used right now, we get it for possible future optimizations.

max_len: usize

The maximum length of the sequence that these T objects can score

parameters: &'a Vec<T>