Trait vcf_rider::rider::CanScoreSequence

pub trait CanScoreSequence {
    fn get_score(&self, pos: usize, sequence: &[u8]) -> f64;

    fn get_length(&self) -> usize;

    fn get_name(&self) -> &str;
}

Our vcf_rider main function will receive a Vec and call it for every T on subsequences of the genomes of the samples doing it only for each variant subsequence once. This trait will need to be able to compute a score on a given sequence, represented by a splice of an array of u8 [TODO] starting for a given position (it is guaranteed by the lib that the used position will be given inside the sequence, i.e. sequence.len() - self.get_length() >= 0).

Required Methods

fn get_score(&self, pos: usize, sequence: &[u8]) -> f64

Returns a score for the given sequence starting at position pos.

Arguments

self - the object with trait CanScoreSequence.

pos - the position in the sequence where the score will be calculated. The check that sequence.len() - self.get_length() >= 0 IS NOT DONE HERE.

sequence- the sequence that needs to be scored, encoded as [ACGTN]-[01234]

fn get_length(&self) -> usize

Returns the length of sequence that this object can score.

Arguments

self - the object with trait CanScoreSequence.

fn get_name(&self) -> &str

Returns the name of this object.

Arguments

self - the object with trait CanScoreSequence.