Module vcf_rider::mutations

Module able to load mutations from a vcf file.

Structs

Coordinate
Struct to represent genomic coordinates. 0 based, end exclusive.

Mutation
Struct used to represent SNPs and indels loaded from vcf files. Suitable only for biallelic polymorphisms.

VcfReader
Struct used to load vcf, it will be used as an Iterator<Mutation>. If accept_phased is false it will panic if there are any unphased mutations, otherwise it will accept them, but beware: sequence where the score will be computed for vcf_rider usually needs to be phased (suppose that two SNPs falls "inside" the same PWM match for the TBA: their phasing will influence the computed score XXX check).

Enums

Position
Enum used to represent the relative position of two Coordinate.