Struct vcf_rider::pwm::PWM

pub struct PWM {
    pub name: String,
    pub ll: Matrix,
    pub llrc: Matrix,
    pub freq: Matrix,
}

Struct representing a whole Positional Weight Matrix.

Fields
name: String

The name of this PWM

ll: Matrix

A Matrix storing likelihoods for this PWM

llrc: Matrix

A Matrix storing reverse complement likelihoods for this PWM. i.e. its elements 0,0 has the likelihood of a T (reverse of A) in the last position of the PWM - used to efficiently score the reverse complement sequence while analyzing sequences going forward.

freq: Matrix

A Matrix storing the original frequencies for nucleotides, these becomes likelihoods when divided by a given sequence background.

Methods

impl PWM

pub fn compute_ll(&mut self, bg: &Vec<f64>)

Given a PWM that has frequencies loaded and some background frequencies it fills ll and ll_rc.

Arguments

bg - a vector with A C G and T nucleotide frequencies to be considered as background for this PWM.